#4 Splash 화면 생성
