+++ now_escape/Loading.js
@@ -1,0 +1,23 @@
+import React from "react";
+import { StatusBar } from "expo-status-bar";
+import { Text, View, StyleSheet } from "react-native";
+
+export default class Loading extends React.Component {
+  render() {
+    return (
+      <View style={styles.container}>
+      <Text>Spalsh</Text>
+      <StatusBar style="auto" />        
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
